requirements.txt
# Core dependencies - compatible with Python 3.13
transformers==4.38.0
torch==2.2.1
scikit-learn==1.4.1.post1
flask==3.0.2
sentence-transformers==2.2.2
numpy==1.26.4
pandas==2.2.0

# YouTube data collection
google-api-python-client==2.115.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
yt-dlp==2024.4.9

# NLP & Text processing
nltk==3.8.1
textblob==0.18.0.post0

# Web & API
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.2
colorama==0.4.6

# Alternative to Rasa for Python 3.13
langchain==0.1.5
openai==1.12.0  # For using GPT models via API
